created separate pglogfileparser class
