+++ src/main/java/com/oltpbenchmark/benchmarks/replay/util/ReplayFileParser.java
@@ -1,0 +1,5 @@
+package com.oltpbenchmark.benchmarks.replay.util;
+
+public class ReplayFileParser {
+    
+}
